Transfer controller  methods as module on concern folder
