+++ app/controllers/messages_controller.rb
@@ -1,4 +1,6 @@
 class MessagesController < ApplicationController
+  include PostMessage
+
   def create
     # このアクションに送られてくるのは新しいquestionなので、questionをnew&saveする。
     @question = Question.new(recipe_id: params[:recipe_id])
@@ -36,28 +38,5 @@ class MessagesController < ApplicationController
     else
       redirect_recipe_when_fail('投稿の削除に失敗しました')
     end
-  end
-
-  private
-  def new_message_params
-    params.require(:message).permit(:context).merge(user_id: current_user.id, question_id: @question.id, is_recipe_user: check_status, is_first: '0')
-  end
-
-  def reply_message_params
-    params.permit(:context, :question_id).merge(user_id: current_user.id, is_recipe_user: check_status, is_first: '1')
-  end
-
-  # messageの投稿者がrecipeの投稿者と同じか判定
-  def check_status
-    if current_user.id == Recipe.find(params[:recipe_id]).user_id
-      return '0'
-    else
-      return '1'
-    end
-  end
-
-  def redirect_recipe_when_fail(message)
-    redirect_to controller: :recipes, action: :show, id: params[:recipe_id]
-    flash[:danger] = message
   end
 end
